requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.3
requests>=2.32.3
beautifulsoup4>=4.12.3
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.4
